chore: store refactor and naming convention
