+++ src/features/cart/Cart.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { getCart } from '@/store/cart/cartStore';
+import { getCart } from '@/store/cart/cart.get';
 import { useQuery } from '@tanstack/react-query';
 import { CartType } from '@/types';
 import RemoveCartItemButton from './ui/RemoveCartItemButton';
