+++ src/features/cart/ui/UpdateCartItem.tsx
@@ -1,4 +1,7 @@
-import { IUpdatedCart, updateCartItemQuantity } from '@/store/cart/cartStore';
+import {
+  IUpdatedCart,
+  updateCartItemQuantity,
+} from '@/store/cart/cart.update-quantity';
 import { UnitsOnCart } from '@prisma/client';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useState } from 'react';
