+++ src/store/cart/cart.add.ts
@@ -1,0 +1,13 @@
+export const postUnitsToCart = async (unitId: string): Promise<Response> => {
+  const data = await fetch('/api/cart/add', {
+    method: 'POST',
+    headers: {
+      'Content-type': 'application/json',
+    },
+    body: JSON.stringify(unitId),
+  })
+    .then((res) => res.json())
+    .catch((err) => err);
+
+  return data;
+};
