+++ src/store/products.get.ts
@@ -1,3 +1,4 @@
+// TODO: fix /types
 import { Product } from '@/types';
 
 export const getProducts = async (): Promise<Product[]> => {
